Added ability to use additional optimizers
To add the ability to use additional optimizers without editing the core loader, I have added the option additionalPlugins to the loader options. Using this, a user can define an additional imagemin plugin to use, along with it's options, and the loader will utilize that optimizer as well (as long as it is installed as a node module).